Moves privateRoute from containers to views
